editoast, front: make paced field optional
editoast
- Adapt models and schemas
- Rework TrainId to better handle the merge of train schedule and paced
  trains

front
- Introduce temporary types of PacedTrains with paced field defined: PacedTrainWithPaced, PacedTrainResponseWithPaced
- Handle a paced train without paced field as a TrainSchedule (single train)
- Fix the conflicts/occupations endpoints and data workflow
- Fix labels when editing a pacedTrain without paced
- Fix RoundTrips suggestions for PacedTrains without paced
- Fix selected items labels to handle PacedTrains without paced
- Fix selectedTrainId process when creating or editing pacedTrains without paced field:
In useAutoSelectTrainId, if the train to select is a pacedTrain without paced, then the id to
select is its raw id, and not the id of its first occurrence. It's the same when editing a train.

The naming may not be ideal, but all this code will be cleaned in the coming weeks.

+++ editoast/schemas/src/fixtures.rs
@@ -149,7 +149,7 @@ pub fn simple_created_exception_with_change_groups(key: &str) -> PacedTrainExcep
 
 pub fn simple_modified_exception_with_change_groups(
     key: &str,
-    occurrence_index: i32,
+    occurrence_index: usize,
 ) -> PacedTrainException {
     let mut exception = simple_created_exception_with_change_groups(key);
     exception.exception_type = ExceptionType::Modified { occurrence_index };
